refactor(price): ValueObject 적용
