+++ src/main/java/lotto/domain/Lottos.java
@@ -3,8 +3,6 @@ package lotto.domain;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static lotto.constant.Constants.LOTTO_PRICE;
-
 public class Lottos {
     private final List<Lotto> lottos;
     private final LottoNumberGenerator lottoNumberGenerator;
@@ -14,10 +12,8 @@ public class Lottos {
         this.lottoNumberGenerator = lottoNumberGenerator;
     }
 
-    public void buyLotto(int price) {
-        checkPrice(price);
-        int lottoCount = price / LOTTO_PRICE;
-        for (int i = 0; i < lottoCount; i++) {
+    public void buyLotto(LottoPrice lottoPrice) {
+        for (int i = 0; i < lottoPrice.lottoCount(); i++) {
             lottos.add(new Lotto(lottoNumberGenerator.lotto()));
         }
     }
@@ -26,12 +22,6 @@ public class Lottos {
         return lottos.stream()
                 .map(lotto -> winningNumbers.match(lotto))
                 .collect(Collectors.toList());
-    }
-
-    private static void checkPrice(int price) {
-        if (LOTTO_PRICE > price) {
-            throw new IllegalArgumentException("로또 구매 비용이 부족합니다.");
-        }
     }
 
     public int quantity() {
